controllers: Remove BaseController
Using inheritance to share few helper methods is not a proper design
and, if we wanted to use dependency injection autowiring, we would
have to pass the dependencies down in every child class constructor.
For now, let’s move the methods to the Authentication helper, which,
while not the best place, having too many responsibilities already,
is at least thematicaly relevant.

Since the BaseController is now empty, we can get rid of it.
